requirements.txt

fastapi>=0.111.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.3
yt-dlp>=2025.1.26
openai-whisper>=20231117
torch>=2.2.0
numpy>=1.26.0
ffmpeg-python>=0.2.0
pydantic>=2.7.0
